+++ src/app/factories/config-initializer.factory.ts
@@ -1,0 +1,5 @@
+import { ApiConfigurationService } from '../services/api-configuration.service';
+
+export const configInitializerFactory =
+  (apiConfigService: ApiConfigurationService) => () =>
+    apiConfigService.loadConfig();
